// frontend/types.ts
export enum Role {
  USER = 'user',
  ADMIN = 'admin',
  MANAGER = 'manager',
  OWNER = 'owner',
}